Reduce size of background image
The background image is nearly 1MiB, despite the fact that this adds
very little to the informational content of the site. This change
reduces the image's size by 90% (it's less than 100KiB now), but
shouldn't change the user experience by much.

I resized it (reducing resolution by a factor of 2 in both dimensions)
and recompressed it using jpeg quality 55 using graphicsmagick:

    gm convert static/img/landing_page_background_img.jpg -resize 960 -quality 55 static/img/landing_page_background_img.smaller.jpg
    mv static/img/landing_page_background_img.smaller.jpg static/img/landing_page_background_img.jpg

This should make the site load much faster on low-bandwidth
connections.