Make compendium URLs more distinct (#30)
As it currently stands, each compendium entry is indexed from zero to the number of entries in the compendium, which has two downsides. First, given a URL, it's very unclear where the URL is pointing to, and second (and more importantly), any update to the compendium will potentially change the corresponding entry number, which breaks links to pages.

This PR gives each page a distinct name based on a slug of the title of the entry, plus a truncated hash of it. There's still a little work that could be done to improve the URLs -- e.g., it's unclear whether appending the entry hash is actually the best way to handle duplicate slugs -- but once this is merged in, most of the problems related to entry links should be fixed.

I've also included some more minor improvements, e.g. making the `generate_compendium.py` script slightly more readable (in my opinion), and stopping printing queries as they're typed with `console.log(...)` (which shouldn't have been committed to the repository in the first place.)

Closes #17.