Reduce likelihood of unreproducible entry digests
As long as an entry's URL is a digest is generated from the json of an
entry, we should try to ensure that the serialization is as
reproducible as possible.  This change makes it independent of the
default parameters for the most obviously flexible parts of
json.dumps, which might change with revisions of the json module)

see also https://reproducible-builds.org/docs/stable-outputs/

This also limits the fields chosen that will produce the hash, at the
suggestion of @kernelmethod.

Addresses (but does not completely close) #34