Switch the default branch from "master" to "main".
Modify the GitHub Actions workflow so that it only runs the staging site
deployment step on pushes to main, rather than pushes to master. This
also updates DEVELOPMENT.md to refer to "main" rather than "master", and
the build_local.sh script to fetch data.bib from the "main" branch of
the GitHub Pages repo.

Closes #41.