Update the Fuse.js version and configuration options.
I've done a few experiments changing around the search parameters used
by Fuse.js, and it seems that (a) turning down the threshold a bit and
(b) making the search location-independent (either by setting a large
"distance" or by setting ignoreLocation to true) seem to fix the
majority of the issues that currently exist with search. For this
commit, I've chosen to tune threshold from 0.3 to 0.1, and I've set
distance to 3000 (so Fuse.js should search the title, authors, and
abstract all up to the first 300 characters of each). It's possible that
some of these parameters will need to be changed further in the future
(especially the distance parameter), but heuristically, these new
options seem to be working pretty well, and don't slow down search too
much.

In addition to changing the Fuse.js parameters, I've also updated it
from v5.1.0 to v6.4.6, which seems to improve search results just a
little more.

This commit is a partial fix for #25, but some additional work will
likely need to be done to make search work correctly.