Check bibtex for duplicate indexes
This is a simple test -- any duplicate bibtex ID will cause all but
one of the entries to be skipped in entries_dict due to the collision.

The goal is to abort the site build if `data.bib` has such a duplicate
entry.

This is still pretty weak: it doesn't test whether `data.bib` is valid
bibtex, it doesn't verify anything else about the entries, but it also
offers a place to do more in-depth consistency/cleanliness checks on
the data scraped from zotero.

Addresses (but does not close) #46.