Upgrade to Vue.js v3.2.12.
This commit upgrades Vue.js to the latest version. The conversion from
v2 -> v3 broke a lot of the search functionality on the site, so search
has been reworked somewhat in order to get it to work with the new
version of Vue.js.

Since I'm slightly more experienced with Vue than I was before, I've
decided to lean a little more on Vue features in order to greatly
simplify the code to render the search results. For instance, instead
of manually deleting every node in the search result list and then
re-inserting them back into the list, I've opted to use Vue's v-for
property on a search-result element to iterate over the list of results.
As a result, the new code is arguably much simpler (and probably more
efficient) than it was previously.